Rust File Manipulation
Got help from GPT to get the ball rolling.

I am going to work with the world files in the native Rust section, then send back and forth data that it reads and writes by way of the UI.

I am actually extremely happy that this is in Rust and not in JS, like I would have wanted a few years ago.