Mojang LevelDB + Tempfile
This is the library that GPT seemed to recommend, I didn't notice it was this one initially, I thought it was the similarly-named std library feature instead. This apparently auto-cleans these files for you after they go out of scope, as of other Rust rules.

https://docs.rs/tempfile/latest/tempfile/

Now this can open other DB types, specifically older worlds that use zlib-based compression.

I want to find a way to log out what the compression is for a given world, that would be great to be able to persist consistently like when making edits, as well as to just make note of it so that we can document it for other usages too. It seems like newer worlds use one kind of compression (uncompressed?), while older ones use zlib for sure (either regular zlib, or zlib-raw [deflate, deflate-raw I think]).

The config for the DB here is from my `player-editor` project, I collected the info to make it from a few different resources. `rusty-leveldb` has it's own example that includes a similar setup to this which was a reference for me to get it working too.