ASCII Only (in concept)
This still doesn't seem to be working properly. I think I need to slim down the character ranges I want to allow, as there are definitely plenty of ASCII characters that aren't neatly renderable.

It was hard to work out how to properly map the errors to only allow for a subset of error cases and change those into when I would want to return a HEX string instead. GPT helped a bit, and my own searching and debugging helped move it along.

https://stackoverflow.com/questions/67290711/rust-error-handling-ive-got-a-utf8error-i-want-to-return-ioresult

Spanish Flyght
Your Pequad