Infinite Scroll DB Keys
This kind of seems to help, but it's still really slow for big worlds. I think I need to debug what data it is passing back and forth.

https://github.com/bvaughn/react-window ⭐️
https://github.com/Lodin/react-vtree (might use this later down the road too)

I realize now that I decided to use React because of the big subset of libraries out there that can make things like this much easier. As I want to get this project up and running in concept before execution, I really want to use the most available tools out there this time around. Not being afraid of the 'status quo' or 'cliche' can be very powerful!