Initial Auto-Build!
This is actually the first release to be real, it seems like when you publish a new GitHub Actions file without any history in the repo, up to the remote, it doesn't run it the very first time, only on subsequent changes. So this will enable that, and hopefully have a successful build and releases for people to download!

LOTS of work needs to be done for this app still, I just need to see that certain things are possible before I make it nice and tidy.