Tauri Conf Version (oops)
Dang, missed another version number. Is there a way to update all of these at once? Kind of a mess, I would have thought that this one would line up with the others automatically, but maybe that's too complex (understandable). I guess 0.1.1 is actually numbered as 0.1.1 on the GitHub Release and app builds. Not ideal, but it's not that big of a deal for now I guess 🤷‍♂️