WorldEditor View
Got more help from specific explanations with GPT, and it laid out a nice way to organize this so that you can render the DB/folder contents in a React kind of setting.

Should the results and prompts of working with GPT be part of the repo? Kind of a neat concept. Especially with prompts themselves.

I really want to make a local backup of all my history with GPT so far sometime. I hope they have a way to bulk download everything like that, I would love to save some of the things I've learned on here. It's one thing to summarize it, another to have it completely as it is, full in-depth. Might actually look into doing just that.