WorldEditor DB Tree View (Partial)
Props issues preventing things from running, it seems to be effect-related.

Found this album last June, I really liked the sound of it, but I wasn't in the mood for it yet. I thought of it tonight, and it feels great.
https://www.youtube.com/watch?v=y0IqiVjYLGw

I think it came back into my head after watching this new video. I really do like and really relate to this sound. It's beautiful in what it represents, and the relief it provides for what we all seem to be going through growing up in this time together.
https://www.youtube.com/watch?v=-TevMSwEuSs

Was also listening to Celebration and Catspoon today, brings me back to remembering how happy music can me. I think I am realizing I was in a hard set of times, and listening to heavy music was representative of that. I'm not sure if it's one way or the other, I think if anything it seems like I gravitate to music that parallels the way I currently am able to feel. Right now, this music seems right. Strapping doesn't feel like how I want to feel right now. I absolutely love it, but I don't want to feel that right now. I think it was starting to feel mundane too, not the music itself, but my mental capacity to connect with it. If you listen to the same things day in day out, then you start to lose less and less of what it was you were grasping at in it. This week seems like a very go with the flow, let go, kind of week. Not worrying about things, just letting them be. Being helpful, gratuitous, and happy. Trying to be positive, seeing what that's like. I think getting stuck in the thought loop can bring me down sometimes, so this week I was trying to do other things, keeping note of what made me feel certain ways, why, and seeing if I could respond or do something different, which might make me feel better than it would otherwise. You fall into doing things a certain way because it's what you know, same with things you tend to connect with. It's okay to be in the status quo, currently I'm trying to embrace that. It's okay to be "cheeky" and or "cliche", don't be so hard on yourself. Just point your compass in what you think is right. If someone says you've gone too far, then adjust direction. You don't have to anticipate every single little thing because you feel like a nuisance to others. You're not. You also don't have to listen to your inner commentary. Just go with your feelings, it's okay. It might be hard to find them in the soup that is usually in the way, but they are there. Just try to stay still, you will see them shine through the debris. They are your light to help you shine brighter. People love to see you, so try to see how you can find you in there!

6w5