Extraction Tree Debug!!
Yay, it's working now!!! The names for things were just all goofed up between Rust and TS. This is the kind of thing I don't want to worry about with this, and it's fairly annoying. It's simple issues like this that can break the whole thing, and it easily could be fixed with explicit type-checking.

Needed fix: Items need to be keyed.