GH Release from Tag
I am learning how it works from over in my player-editor project. This part is needed to make a GH release from the pre-existing tag. I guess I thought they were the same thing, oops. A release references an existing tag (or could be a new one altogether, solely for the release itself), but a tag is not in of itself a release.

Ok this is the new and hopefully final build of 0.1.2.