GH Release from Tag Revert
Can't seem to get the config for this working yet, so I'm not going to get hung up on it not working in the moment. Just going to go with what works, in absolute terms. This is usually what holds me up in getting things done most of the time, that I want to get things done in a specific way, and that makes me keep trying a bunch of different ways to get it set up instead of moving on to the next things I should be focusing on (should being opinionated, it's the next thing that would allow the project to progress as a whole. This is the goal I am trying to remind myself of moreso with this. If things are messy, I'd rather have them working, then not as complete, but perfect).