Tagged Extracted Entries
Relentless, syl

Learning about how serde tags work. Looks like the enum itself is what makes it work, so having a root item doesn't add the tag, it has to be referenced with the enum itself.

Still trying to allow myself to move fast rather than get it exactly right to start with, it's really a mental challenge not to at least *try* to do it perfectly the first time.

https://serde.rs/enum-representations.html
https://serde.rs/attr-rename.html
https://serde.rs/container-attrs.html
https://www.reddit.com/r/rust/comments/186sfd8/serde_today_i_learned_that_the_tag_attribute/
https://www.reddit.com/r/learnrust/comments/1c5jlbd/how_to_convert_to_snake_case_with_serde/