Invoke Abstraction + Better Error Notifications
Now error information is persisted to the console. I also intend to solely use `invoke()` only through dedicated functions, as that will better break up where certain Tauri features are implemented. I want to be able to test the frontend in a normal browser, without needing to have all of the app's features available. I also am doing this to mitigate the usage of generics leak for intended types when using `invoke()` itself. Since it is generic (which is good for proper annotations without assertions), this has the possibility to fall out of sync with the actual usage. I want to ensure that these weak transversal points between the frontend and the backend are limited to small endpoints, only a few areas.