Undefined State Removal
This fix was from a while back in the process, when things weren't being assigned correctly for one reason or another. I think it had to do with memoization, and that seems to be fixed for a bit now.

Symbolic Fox revisit, love that album, it was my first Marco deep dive, and it is still absolutely one of my favorites to this day :)