MCWorld Module Split
Going to start implementing more features that work alongside the world data itself, this will allow for adding new runtime features, as now the database might be held as a reference and opened for a longer period of time, since I no longer will need it solely for opening files. I want to persist the database throughout the duration of editing the world itself. Once the user closes the app or opens a different world, the database should be closed.