MCWorld Flattening Pt. 1
Going to abstract away parts of this current function, as it is much too big to be maintainable. I'm surprised it's this simple at this point. This reminds me of how STE came about, except that instead of refactoring early like I'm about to do here (before adding more features on top), I just kept building more and more and more on top of the existing ball of twine. That's how you get something you don't want to work on anymore haha. I love that Rust doesn't have the global scope though haha. I also really like that you have to be very intentional with your programming when it has to be done in separate segments like this though. Not having the JS/UI available really make it explicit how your program should be able to interact with that outside world. It is great. Understanding that limitations are what guide you is awesome.