Open MCWorld Rename
Going to change the name of this function, since this is how the main part of the app is intended to be pulled from, as of yet. The scope of what this function implements is about to change though, as it will have to maintain the DB instance itself for a bit, which it currently disposes of directly after opening the world file.