read_archive()
Pointers/mutable pointers are interesting, still trying to understand things. I get the rules, but I don't get what one might mean over the other, depending on whether you're using a pointer, mutable pointer, or pointer to mutable data. Something like that. This was the error I was running into before a little change to this code:

```
cannot borrow data in a `&` reference as mutable
cannot borrow as mutable
```