read_archive() Side Effect Removal
Now this returns the data itself, rather than modifying the vec passed into it. Might as well/makes more sense to wrap the mutability as well, that's only to make the vec itself. Didn't realize just how great this is in Rust, that mutability is just a runtime-enforcement. I thought this changed the value itself, meaning once mutable, it can only be mutable. This is not true! Awesome! And the fact that you can return mutable data from a function as a solid immutable one, that's outstanding. Loving this, makes even more sense now.

Looks like this has to revert the pointer changes from the previous realization, I think it has to do with referencing regular values in the same scope. It doesn't make sense to have a mutable pointer inside your function locally, you'd rather do that only for functions that don't return those, I think, or something like that.