read_entries() Side Effect Removal
Man I wish TypeScript had type checking for this kind of thing, it is outstanding. I love it. Maybe it's just a programming mindset to follow, I could see it just being that too. Having explicit references/ownership is the part to make note of though, it's a whole new way to look at your codebase.