MCWorldViewer
Going to make a new way to open things, rather than passing in the file contents with/from the GUI itself and a regular file input, I'm going to use a Tauri file picker dialog, then pass that path to the Rust backend, to have it do the file fetching instead. I think this will make things easier down the road to make this a solely CLI-based app too, without the need of a Tauri frontend all of the time. I absolutely want that too, in the later event of making a converter for world data. That will be a separate iteration though, I want to have this working here first. This is kind of like Dovetail, where it's a way for me to learn how things work, and how to implement them.