Test World Demos
Going to set up opening plain world databases too, so now you will be able to open both `.mcworld` and folder-based worlds. This is still solely focused on Bedrock parsing.

(I committed this originally before the previous update, but I moved it to here in the commit history because I wanted to make it clear it is meant to be related to the next coming changes, rather than the ones in the build)

These are the same files from my player-editor project, and Player-Migrator originally.