MCWorld Extractions and de-"cross-scoping"
Realizing where in the root function that things could have been broken apart better. They didn't need to be in the same scope, so taking them out and into their own functions better explains what needs what, and what doesn't. This is what abstractions are for, not always just to dedupe things. This makes so much more sense! I think I have understood this for a bit now, but not outright to this extent. Breaking things apart into separate functions also helps explain why something is single-use, and in what way. Having only certain variables defined in the scope makes it much simpler to see what it does, and it doesn't allow you to add accidental cross-polination, if that makes sense. One feature shouldn't need to function based on another one being in-scope. Both features should work in tandem, and individually. This is what brought and still brings STE to it's knees, everything is hairily nested together like a big ball of twine. It all works, but picking one thing out to reuse again later, heck no. It shouldn't be that way, and that's the thing I'm realizing now. It's kind of like the difference in repairability for a SOC architecture, vs individual components you can repair. Except in programming, you don't get any benefits from having everything so tightly-coupled. Maybe just arbitrary things like "small amount of functions", but that certainly doesn't help with anything.