open_db_dir() Setup
This function doesn't do anything yet, I need to differentiate it still.

I am also realizing how pre-abstracting your code before it is complete is what tends to lead to chaos in the codebase, so I think leaving things in the same module initially is okay when you are trying to find out what the similarities are between the implementations, and what needs the same types and what doesn't.

https://www.youtube.com/watch?v=bZE7npzPrY0