convertToNode() Flattening
Moving this over to a mapper predicate rather than a mapper itself. This allows for more composability :)

Took a break an hour or two ago, and happened to be watching these videos, and it lines up well with this actually hehe.

https://www.reddit.com/r/rust/comments/1d3hvhw/why_rust_doesnt_have_classes/
https://www.youtube.com/watch?v=hxGOiiR9ZKg
https://www.youtube.com/watch?v=nuML9SmdbJ4

Been working on artwork for my earlier projects, which weren't 100% complete yet. Nice to go back and listen to what I was writing over the last few years. It definitely has it's waves of inspiration and concepts. I'm at the stage of things where I want to move back into the next phase again, I was in this same feeling last September. So it's every few months or so. Back in fall then, I was stuck in a metal/chaos phase, and now I'm stuck in a relaxed ambient phase. I want to make more cohesive songs, and less explorative. I tend to seem keep writing 'aimless' songs though. It might be what I need right now though. And I'm trying to accept when I need to write what I need to write, in that moment. I've learned and try to remind myself, forcing yourself to write something you aren't feeling is not what you want to do. So, it's okay. I'm reminding myself right now, it's okay what I'm doing now, this time will pass, and the next one will arrive when it's time. You need this now, because it will allow you to get to where you want to be ⭐️