Handler Declaration Fix
This seems to be the issue, it was how I was generating the handlers for the methods. The issue linked that the parameters could be the issue, but it wasn't that. Still great info for things though too, didn't completely realize that the arguments would be passed through with Serde directly like that, makes sense why the casing would change though.

https://v1.tauri.app/v1/guides/features/command/
https://v2.tauri.app/develop/calling-rust/
https://stackoverflow.com/questions/78432685/why-does-tauri-modify-the-parameter-names-of-invoked-functions