Node Entry Aliases
Moving over to these because I want to flatten the usage of `ExtractionEntry` & friends to only use the ones needed by the renderer, rather than mapping these individually. This is the setup for that. I want the Rust types to be exactly what the UI needs to be rendered, without needing to map things.