macOS Minimum Build Version Fix
Yay! This was annoying, I found a fix previously, and now this is the proper fix, it seems like, I hope. Found it in a comment someone replied with, to an issue I was subscribed to.

https://github.com/tauri-apps/tauri/issues/11577#issuecomment-2869396570
https://yuexun.me/how-to-make-your-tauri-dev-faster/

My original solution seemed to be working, but this is more accurate to what I would have hoped it to be anyways, since it shouldn't be the fix for the development build to just raise it to a higher minimum build version. That's not the issue. It's just that they were out of sync.